Remove Timer.cpp/h. Move various collections into Common/Data/Collections.
