+++ Common/Common.vcxproj.filters
@@ -7,18 +7,15 @@
     <ClInclude Include="CommonTypes.h" />
     <ClInclude Include="ConsoleListener.h" />
     <ClInclude Include="CPUDetect.h" />
-    <ClInclude Include="FixedSizeQueue.h" />
     <ClInclude Include="Log.h" />
     <ClInclude Include="LogManager.h" />
     <ClInclude Include="MemArena.h" />
     <ClInclude Include="MemoryUtil.h" />
     <ClInclude Include="StringUtils.h" />
     <ClInclude Include="Thunk.h" />
-    <ClInclude Include="Timer.h" />
     <ClInclude Include="x64Analyzer.h" />
     <ClInclude Include="x64Emitter.h" />
     <ClInclude Include="ArmEmitter.h" />
-    <ClInclude Include="ThreadPools.h" />
     <ClInclude Include="Crypto\md5.h">
       <Filter>Crypto</Filter>
     </ClInclude>
@@ -33,12 +30,10 @@
     <ClInclude Include="MipsEmitter.h" />
     <ClInclude Include="Arm64Emitter.h" />
     <ClInclude Include="ArmCommon.h" />
-    <ClInclude Include="MathUtil.h" />
     <ClInclude Include="BitSet.h" />
     <ClInclude Include="CodeBlock.h" />
     <ClInclude Include="ColorConv.h" />
     <ClInclude Include="ColorConvNEON.h" />
-    <ClInclude Include="ThreadSafeList.h" />
     <ClInclude Include="GL\GLInterface\EGL.h">
       <Filter>GL\GLInterface</Filter>
     </ClInclude>
@@ -51,7 +46,6 @@
     <ClInclude Include="GraphicsContext.h" />
     <ClInclude Include="DbgNew.h" />
     <ClInclude Include="OSVersion.h" />
-    <ClInclude Include="Hashmaps.h" />
     <ClInclude Include="BitScan.h" />
     <ClInclude Include="MakeUnique.h" />
     <ClInclude Include="ExceptionHandlerSetup.h" />
@@ -372,6 +366,15 @@
     <ClInclude Include="Render\Text\draw_text_win.h">
       <Filter>Render\Text</Filter>
     </ClInclude>
+    <ClInclude Include="Data\Collections\FixedSizeQueue.h">
+      <Filter>Data\Collections</Filter>
+    </ClInclude>
+    <ClInclude Include="Data\Collections\Hashmaps.h">
+      <Filter>Data\Collections</Filter>
+    </ClInclude>
+    <ClInclude Include="Data\Collections\ThreadSafeList.h">
+      <Filter>Data\Collections</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="ABI.cpp" />
@@ -381,12 +384,10 @@
     <ClCompile Include="MemoryUtil.cpp" />
     <ClCompile Include="StringUtils.cpp" />
     <ClCompile Include="Thunk.cpp" />
-    <ClCompile Include="Timer.cpp" />
     <ClCompile Include="x64Analyzer.cpp" />
     <ClCompile Include="x64Emitter.cpp" />
     <ClCompile Include="ArmEmitter.cpp" />
     <ClCompile Include="ArmCPUDetect.cpp" />
-    <ClCompile Include="ThreadPools.cpp" />
     <ClCompile Include="Crypto\md5.cpp">
       <Filter>Crypto</Filter>
     </ClCompile>
