+++ Common/TimeUtil.h
@@ -5,3 +5,5 @@ double time_now_d();
 
 // Sleep. Does not necessarily have millisecond granularity, especially on Windows.
 void sleep_ms(int ms);
+
+void GetTimeFormatted(char formattedTime[13]);
