+++ Core/HW/StereoResampler.cpp
@@ -36,7 +36,7 @@
 #include <atomic>
 
 #include "Common/System/System.h"
-#include "Common/MathUtil.h"
+#include "Common/Math/math_util.h"
 #include "Common/Serialize/Serializer.h"
 #include "Common/Log.h"
 #include "Common/TimeUtil.h"
