+++ GPU/D3D11/DrawEngineD3D11.h
@@ -21,7 +21,7 @@
 #include <d3d11.h>
 #include <d3d11_1.h>
 
-#include "Common/Hashmaps.h"
+#include "Common/Data/Collections/Hashmaps.h"
 #include "GPU/GPUState.h"
 #include "GPU/Common/GPUDebugInterface.h"
 #include "GPU/Common/IndexGenerator.h"
