+++ GPU/D3D11/TextureScalerD3D11.cpp
@@ -19,7 +19,7 @@
 
 #include <d3d11.h>
 #include "Common/ColorConv.h"
-#include "Common/ThreadPools.h"
+#include "Core/ThreadPools.h"
 #include "GPU/Common/TextureScalerCommon.h"
 #include "GPU/D3D11/TextureScalerD3D11.h"
 #include "GPU/D3D11/GPU_D3D11.h"
