+++ GPU/Directx9/DrawEngineDX9.h
@@ -19,7 +19,7 @@
 
 #include <d3d9.h>
 
-#include "Common/Hashmaps.h"
+#include "Common/Data/Collections/Hashmaps.h"
 #include "GPU/GPUState.h"
 #include "GPU/Common/GPUDebugInterface.h"
 #include "GPU/Common/IndexGenerator.h"
