+++ GPU/GLES/DrawEngineGLES.h
@@ -17,8 +17,9 @@
 
 #pragma once
 
-#include <Common/Hashmaps.h>
-#include <unordered_map>
+#include "Common/Data/Collections/Hashmaps.h"
+#include "Common/GPU/OpenGL/GLCommon.h"
+#include "Common/GPU/OpenGL/GLRenderManager.h"
 
 #include "GPU/GPUState.h"
 #include "GPU/Common/GPUDebugInterface.h"
@@ -27,8 +28,6 @@
 #include "GPU/Common/DrawEngineCommon.h"
 #include "GPU/Common/GPUStateUtils.h"
 #include "GPU/GLES/FragmentShaderGeneratorGLES.h"
-#include "Common/GPU/OpenGL/GLCommon.h"
-#include "Common/GPU/OpenGL/GLRenderManager.h"
 
 class LinkedShader;
 class ShaderManagerGLES;
