+++ GPU/Software/Rasterizer.cpp
@@ -20,7 +20,7 @@
 
 #include "Common/Profiler/Profiler.h"
 
-#include "Common/ThreadPools.h"
+#include "Core/ThreadPools.h"
 #include "Common/ColorConv.h"
 #include "Core/Config.h"
 #include "Core/MemMap.h"
