+++ GPU/Vulkan/ShaderManagerVulkan.h
@@ -20,7 +20,7 @@
 #include <cstdio>
 #include <cstdint>
 
-#include "Common/Hashmaps.h"
+#include "Common/Data/Collections/Hashmaps.h"
 #include "Common/GPU/Vulkan/VulkanMemory.h"
 #include "GPU/Common/ShaderCommon.h"
 #include "GPU/Common/ShaderId.h"
