+++ GPU/Vulkan/TextureCacheVulkan.h
@@ -19,7 +19,7 @@
 
 #include <map>
 
-#include "Common/Hashmaps.h"
+#include "Common/Data/Collections/Hashmaps.h"
 #include "GPU/GPUInterface.h"
 #include "GPU/GPUState.h"
 #include "Common/GPU/Vulkan/VulkanContext.h"
