+++ GPU/Vulkan/TextureScalerVulkan.cpp
@@ -21,7 +21,7 @@
 #include "Common/GPU/Vulkan/VulkanContext.h"
 #include "Common/ColorConv.h"
 #include "Common/Log.h"
-#include "Common/ThreadPools.h"
+#include "Core/ThreadPools.h"
 #include "GPU/Common/TextureScalerCommon.h"
 #include "GPU/Vulkan/TextureScalerVulkan.h"
