+++ android/jni/Android.mk
@@ -300,8 +300,6 @@ EXEC_AND_LIB_FILES := \
   $(SRC)/Common/MemoryUtil.cpp \
   $(SRC)/Common/StringUtils.cpp \
   $(SRC)/Common/SysError.cpp \
-  $(SRC)/Common/ThreadPools.cpp \
-  $(SRC)/Common/Timer.cpp \
   $(SRC)/Common/TimeUtil.cpp \
   $(SRC)/GPU/Math3D.cpp \
   $(SRC)/GPU/GPU.cpp \
@@ -395,6 +393,7 @@ EXEC_AND_LIB_FILES := \
   $(SRC)/Core/Screenshot.cpp \
   $(SRC)/Core/System.cpp \
   $(SRC)/Core/TextureReplacer.cpp \
+  $(SRC)/Core/ThreadPools.cpp \
   $(SRC)/Core/WebServer.cpp \
   $(SRC)/Core/Debugger/Breakpoints.cpp \
   $(SRC)/Core/Debugger/DisassemblyManager.cpp \
